feat!: Merge `c2pa-status-tracker` crate into `c2pa` (#1115)
