+++ internal/status-tracker/README.md
@@ -1,5 +1,9 @@
-# C2PA status tracking implementation
+# `c2pa-status-tracker` crate is DISCONTINUED
 
-[![CI](https://github.com/contentauth/c2pa-rs/actions/workflows/ci.yml/badge.svg)](https://github.com/contentauth/c2pa-rs/actions/workflows/ci.yml) [![Latest Version](https://img.shields.io/crates/v/c2pa-rs-status-tracker.svg)](https://crates.io/crates/c2pa-rs-status-tracker) [![docs.rs](https://img.shields.io/docsrs/c2pa-rs-status-tracker)](https://docs.rs/c2pa-rs-status-tracker/) [![![codecov](https://codecov.io/gh/contentauth/c2pa-rs/branch/main/graph/badge.svg?token=YVHWI19EGN)](https://codecov.io/gh/contentauth/c2pa-rs)
+This crate has been merged into the [`c2pa` crate](https://crates.io/crates/c2pa). It will no longer be maintained or published as a standalone crate.
 
-This crate supports the implementation of `c2pa-rs` and other related crates and is not intended as a general-purpose public API.
+For the most part, all public APIs can be remapped as follows:
+
+`c2pa_status_tracker::xxx` -> `c2pa::status_tracker::xxx`
+
+There will be no further releases of this crate.
