--- internal/status-tracker/rustfmt.toml
@@ -1,17 +1,0 @@
-# IMPORTANT: PR validation uses nightly version of
-# rustfmt.
-
-# Invoke by using:
-#
-# rustup toolchain add nightly
-# cargo +nightly fmt
-
-blank_lines_upper_bound = 1
-empty_item_single_line = true
-format_code_in_doc_comments = true
-group_imports = "StdExternalCrate"
-hex_literal_case = "Lower"
-imports_granularity = "Crate"
-reorder_impl_items = true
-unstable_features = true
-wrap_comments = true
