+++ sdk/src/assertions/box_hash.rs
@@ -13,7 +13,6 @@
 
 use std::{fs::File, io::Cursor, path::*};
 
-use c2pa_status_tracker::validation_codes::ASSERTION_BOXHASH_UNKNOWN;
 use serde::{Deserialize, Serialize};
 use serde_bytes::ByteBuf;
 
@@ -23,6 +22,7 @@ use crate::{
     asset_io::{AssetBoxHash, CAIRead},
     error::{Error, Result},
     utils::hash_utils::{hash_stream_by_alg, verify_stream_by_alg, HashRange},
+    validation_results::validation_codes::ASSERTION_BOXHASH_UNKNOWN,
 };
 
 const ASSERTION_CREATION_VERSION: usize = 1;
