+++ sdk/src/identity/claim_aggregation/ica_signature_verifier.rs
@@ -12,7 +12,6 @@
 // each license.
 
 use async_trait::async_trait;
-use c2pa_status_tracker::{log_current_item, StatusTracker};
 use chrono::{DateTime, Utc};
 use coset::{CoseSign1, RegisteredLabelWithPrivate, TaggedCborSerializable};
 
@@ -33,6 +32,8 @@ use crate::{
         },
         SignatureVerifier, SignerPayload, ValidationError,
     },
+    log_current_item,
+    status_tracker::StatusTracker,
     HashedUri,
 };
