+++ sdk/src/identity/identity_assertion/signer_payload.rs
@@ -13,11 +13,13 @@
 
 use std::{collections::HashSet, fmt::Debug, sync::LazyLock};
 
-use c2pa_status_tracker::{log_current_item, StatusTracker};
 use regex::Regex;
 use serde::{Deserialize, Serialize};
 
-use crate::{dynamic_assertion::PartialClaim, identity::ValidationError, HashedUri, Manifest};
+use crate::{
+    dynamic_assertion::PartialClaim, identity::ValidationError, log_current_item,
+    status_tracker::StatusTracker, HashedUri, Manifest,
+};
 
 /// A set of _referenced assertions_ and other related data, known overall as
 /// the **signer payload.** This binding **SHOULD** generally be construed as
