+++ sdk/src/identity/tests/claim_aggregation/validation.rs
@@ -19,7 +19,6 @@
 
 use std::io::Cursor;
 
-use c2pa_status_tracker::{LogKind, StatusTracker};
 #[cfg(all(target_arch = "wasm32", not(target_os = "wasi")))]
 use wasm_bindgen_test::wasm_bindgen_test;
 
@@ -29,6 +28,7 @@ use crate::{
         tests::fixtures::claim_aggregation::ica_credential_example,
         IdentityAssertion, ValidationError,
     },
+    status_tracker::{LogKind, StatusTracker},
     Reader,
 };
