+++ sdk/src/identity/tests/fixtures/naive_credential_holder.rs
@@ -23,13 +23,15 @@
 use std::fmt::Debug;
 
 use async_trait::async_trait;
-use c2pa_status_tracker::StatusTracker;
 use serde::Serialize;
 
-use crate::identity::{
-    builder::{AsyncCredentialHolder, CredentialHolder, IdentityBuilderError},
-    identity_assertion::signature_verifier::ToCredentialSummary,
-    SignatureVerifier, SignerPayload, ValidationError,
+use crate::{
+    identity::{
+        builder::{AsyncCredentialHolder, CredentialHolder, IdentityBuilderError},
+        identity_assertion::signature_verifier::ToCredentialSummary,
+        SignatureVerifier, SignerPayload, ValidationError,
+    },
+    status_tracker::StatusTracker,
 };
 
 pub(crate) struct NaiveCredentialHolder {}
