+++ sdk/src/identity/tests/validation_method/continue_when_possible.rs
@@ -16,13 +16,13 @@
 
 use std::io::Cursor;
 
-use c2pa_status_tracker::{LogKind, StatusTracker};
 #[cfg(all(target_arch = "wasm32", not(target_os = "wasi")))]
 use wasm_bindgen_test::wasm_bindgen_test;
 
 use crate::{
     crypto::raw_signature::SigningAlg,
     identity::{x509::X509SignatureVerifier, IdentityAssertion},
+    status_tracker::{LogKind, StatusTracker},
     Reader,
 };
 
@@ -407,7 +407,6 @@ async fn no_hard_binding() {
 mod invalid_sig_type {
     use std::io::Cursor;
 
-    use c2pa_status_tracker::{LogKind, StatusTracker};
     #[cfg(all(target_arch = "wasm32", not(target_os = "wasi")))]
     use wasm_bindgen_test::wasm_bindgen_test;
 
@@ -415,6 +414,7 @@ mod invalid_sig_type {
         identity::{
             claim_aggregation::IcaSignatureVerifier, x509::X509SignatureVerifier, IdentityAssertion,
         },
+        status_tracker::{LogKind, StatusTracker},
         Reader,
     };
