+++ sdk/src/identity/x509/async_x509_credential_holder.rs
@@ -104,7 +104,6 @@ mod tests {
 
     use std::io::{Cursor, Seek};
 
-    use c2pa_status_tracker::StatusTracker;
     #[cfg(all(target_arch = "wasm32", not(target_os = "wasi")))]
     use wasm_bindgen_test::wasm_bindgen_test;
 
@@ -116,6 +115,7 @@ mod tests {
             x509::{AsyncX509CredentialHolder, X509SignatureVerifier},
             IdentityAssertion,
         },
+        status_tracker::StatusTracker,
         Builder, Reader, SigningAlg,
     };
