+++ sdk/src/identity/x509/x509_credential_holder.rs
@@ -76,7 +76,6 @@ mod tests {
 
     use std::io::{Cursor, Seek};
 
-    use c2pa_status_tracker::StatusTracker;
     #[cfg(all(target_arch = "wasm32", not(target_os = "wasi")))]
     use wasm_bindgen_test::wasm_bindgen_test;
 
@@ -88,6 +87,7 @@ mod tests {
             x509::{X509CredentialHolder, X509SignatureVerifier},
             IdentityAssertion,
         },
+        status_tracker::StatusTracker,
         Builder, Reader, SigningAlg,
     };
