+++ sdk/src/identity/x509/x509_signature_verifier.rs
@@ -12,7 +12,6 @@
 // each license.
 
 use async_trait::async_trait;
-use c2pa_status_tracker::{log_current_item, StatusTracker};
 use coset::CoseSign1;
 use serde::Serialize;
 
@@ -25,6 +24,8 @@ use crate::{
         identity_assertion::signature_verifier::ToCredentialSummary, SignatureVerifier,
         SignerPayload, ValidationError,
     },
+    log_current_item,
+    status_tracker::StatusTracker,
 };
 
 /// An implementation of [`SignatureVerifier`] that supports COSE signatures
